Added basic loan calc logic

+++ CarLoan/CarLoan/LoanQuote.cs
@@ -3,8 +3,8 @@ namespace CarLoan
 {
     public class LoanQuote
     {
-        public float rate { get; set; }
-        public float payment { get; set; }
+        public double rate { get; set; }
+        public double payment { get; set; }
         public int term { get; set; }
 
         public LoanQuote()
